Run init scripts from user data
Instead of being pre-baked into the AMI by packer, the supervisor config
that runs the init scripts is now generated by a separate script that is
invoked by the instance's user data. This is a prerequisite for making
the network multi-region (Issue #2), as the vault address will need to
be filled into the user data by terraform.