Move VAULT_ADDR completely into terraform
Move the declaration of the VAULT_ADDR environment variable into user data
with a DNS and port filled in by terraform. This makes terraform the sole
source of the vault address, which will enable us to fill it in with the
DNS of a load balancer created by terraform as part of multi-region networks
in Issue #2.