Put vault cluster behind load balancer
As a prerequisite for multi-region networks (Issue #2), the vault cluster needs
to be behind a load balancer so it can be accessed cross region, since consul
cannot be used for service discovery cross region.

In order to do this, the cert tool needed to be pulled into the terraform
workflow, since we need the load balancer DNS before we can generate certificates.

The load balancer is currently internal. Eventually it will be external in
the multi-region case.