Switch to public DNS for inter-node communication
This changes the init-quorum and init-bootnode scripts to use the public
DNS name instead of the private IP to communicate with other nodes. The
public DNS resolves to the private IP from within the VPC and the public
IP from outside the VPC, meaning we can use it for all communications.

Enode addresses must still use IP, so bootnodes currently still advertise
their enode address using their private IP.

This is a prerequisite for multi-region networks (Issue #2)