Separate quorum-vault module from quorum-cluster-region module
This extracts the quorum-vault module from the quorum-cluster-region module
such that values previously retrieved from the module are retrieved from input
variables instead. A quorum-cluster module now uses both modules side by side.

This is part of making a multi-region network (Issue #2), as it will enable us
to duplicate the regional part of the network for each region and have them share
a common vault cluster.