Add move-vault-data.sh script
This allows a user authenticated with quorum-root or root permissions to 
move the data from one vault "slot" to another. The original vault data 
being replaced will be retired to a "graveyard" and the new data will be 
written in its place. This will cause the instance in the "to" slot to 
reboot as a different instance.